requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
GeoAlchemy2==0.12.5
greenlet==1.1.3.post0
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
psycopg2==2.9.4
pyparsing==3.0.9
python-dotenv==0.21.0
Shapely==1.8.5.post1
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.9.0
